==> Folder3_Within-Case/0_master_distance.ipynb <==
import papermill as pm
import os
import shutil

countrycode = 'default'
countryname = 'default'

# List of parameter sets
parameter_sets = [
    {'countrycode': 'AGO', 'countryname': 'Angola'},
    {'countrycode': 'BDI', 'countryname': 'Burundi'},
    {'countrycode': 'BFA', 'countryname': 'Burkina Faso'},
    {'countrycode': 'CMR', 'countryname': 'Cameroon'},
    {'countrycode': 'CAF', 'countryname': 'Central African Republic'},
    {'countrycode': 'COD', 'countryname': 'Democratic Republic of the Congo'},
    {'countrycode': 'COG', 'countryname': 'Congo'},
    {'countrycode': 'ERI', 'countryname': 'Eritrea'},
    {'countrycode': 'ETH', 'countryname': 'Ethiopia'},
    {'countrycode': 'KEN', 'countryname': 'Kenya'},
    {'countrycode': 'MLI', 'countryname': 'Mali'},
    {'countrycode': 'MOZ', 'countryname': 'Mozambique'},
    {'countrycode': 'NER', 'countryname': 'Niger'},
    {'countrycode': 'NGA', 'countryname': 'Nigeria'},
    {'countrycode': 'RWA', 'countryname': 'Rwanda'},
    {'countrycode': 'SDN', 'countryname': 'Sudan'},
    {'countrycode': 'SOM', 'countryname': 'Somalia'},
    {'countrycode': 'SSD', 'countryname': 'South Sudan'},
    {'countrycode': 'TCD', 'countryname': 'Chad'},
    {'countrycode': 'TZA', 'countryname': 'Tanzania'},
    {'countrycode': 'UGA', 'countryname': 'Uganda'}
]

# Loop through each parameter set and run the parametrized notebook
for params in parameter_sets:
    output_notebook = f'output_{params["countrycode"]}.ipynb'
    try:
        pm.execute_notebook(
            '1_distance-parametrized.ipynb',  # Input notebook
            output_notebook,                  # Output notebook
            parameters=params
        )
        # Define the destination path
        base_path = os.getcwd()
        destination_folder = os.path.join(base_path, params['countrycode'])
        if not os.path.exists(destination_folder):
            os.makedirs(destination_folder)
        # Move the output notebook to the destination folder
        shutil.move(output_notebook, os.path.join(destination_folder, output_notebook))
    except Exception as e:
        print(f"Error encountered for {params['countrycode']}: {e}")
        continue


# Execute additional notebooks
additional_notebooks = [
    '2_add-in-capital-state.ipynb',
    '3_create-overview_figure4.ipynb',
    '4_logistic_regression_figure5.ipynb'
]

for notebook in additional_notebooks:
    output_notebook = f'output_{notebook}'
    try:
        pm.execute_notebook(
            notebook,          # Input notebook
            output_notebook    # Output notebook
        )
        # Define the destination path
        base_path = os.getcwd()
        destination_folder = os.path.join(base_path, 'additional_notebooks')
        if not os.path.exists(destination_folder):
            os.makedirs(destination_folder)
        # Move the output notebook to the destination folder
        shutil.move(output_notebook, os.path.join(destination_folder, output_notebook))
    except Exception as e:
        print(f"Error encountered for {notebook}: {e}")
        continue

print("All done.")
