==> Folder3_Within-Case/2_add-in-capital-state.ipynb <==
# Add 'is in capital'

import os
import pandas as pd
# List of country codes
#country_codes = ['AGO', 'BDI', 'BFA', 'CMR', 'CAF', 'TCD', 'COG', 'COD', 'ERI', 'ETH', 'KEN', 'MLI', 'MOZ', 'NER', 'NGA', 'RWA', 'SDN', 'SOM', 'SSD', 'TZA', 'UGA']
country_codes = ['COD']
# Iterate through each country code
for country_code in country_codes:
    # Define the subfolder path
    subfolder_path = os.path.join(os.getcwd(), country_code)
    
    # Define the file paths
    state_capital_file = os.path.join(subfolder_path, f'state_capital_list_{country_code}_points.csv')
    grid_file = os.path.join(subfolder_path, f'{country_code}_grid3_with_ntl_percentage.csv')
    
    # Read the state capital file
    state_capital_df = pd.read_csv(state_capital_file)
    
    # Find the state with the capital
    state_with_capital = state_capital_df[state_capital_df['is_capital_state'] == 1]['state'].values[0]
    
    # Read the grid file
    grid_df = pd.read_csv(grid_file)
    
    # Create the 'in_state_capital' column
    grid_df['in_state_capital'] = grid_df['adm1_name'].apply(lambda x: 1 if x == state_with_capital else 0)
    
    # Print the total number of rows
    total_rows = len(grid_df)
    print(f"Total number of rows for {country_code}: {total_rows}")
    
    # Print the portion of rows where 'in_state_capital' is 1
    portion_in_state_capital = grid_df['in_state_capital'].mean()
    print(f"Portion of rows where 'in_state_capital' is 1 for {country_code}: {portion_in_state_capital:.2%}")
    
    # Define the new file path
    newfile = os.path.join(subfolder_path, f'{country_code}_grid3_with_ntl_percentage_complete.csv')
    # Save the updated grid file (optional)
    grid_df.to_csv(newfile, index=False)
    
    # Delete the original grid file
    os.remove(grid_file)

print("Column is_in_capital created")
