==> Folder3_Within-Case/3_create-overview_figure4.ipynb <==
# Create overview of key stats for figure 4

import pandas as pd
import os

# List of country tuples
countries = [
    ('AGO', 'Angola'),
    ('BFA', 'Burkina Faso'),
    ('BDI', 'Burundi'),
    ('CMR', 'Cameroon'),
    ('CAF', 'Central African Republic'),
    ('TCD', 'Chad'),
    ('COG', 'Congo'),
    ('COD', 'Democratic Republic of the Congo'),
    ('ERI', 'Eritrea'),
    ('ETH', 'Ethiopia'),
    ('KEN', 'Kenya'),
    ('MLI', 'Mali'),
    ('MOZ', 'Mozambique'),
    ('NER', 'Niger'),
    ('NGA', 'Nigeria'),
    ('RWA', 'Rwanda'),
    ('SDN', 'Sudan'),
    ('SOM', 'Somalia'),
    ('SSD', 'South Sudan'),
    ('TZA', 'Tanzania'),
    ('UGA', 'Uganda')
]

# Initialize the dataframe
df_figure4 = pd.DataFrame(columns=['countryname', 'built-perc_mon', 'built-perc_any', 'built_median_any', 'small-perc_mon', 'small_perc_any', 'small_median_any'])

# Iterate through each country tuple
for countrycode, countryname in countries:
    # Define the folder path
    folder_path = os.path.join(countrycode)
    
    # Open 'overview_key_stats.csv' and extract the required values
    overview_stats_path = os.path.join(folder_path, 'overview_key_stats.csv')
    overview_df = pd.read_csv(overview_stats_path)
    
    built_perc_mon = overview_df[(overview_df['type'] == 'Built-up') & (overview_df['what'] == 'Proportion monitorable')]['value'].values[0]
    built_perc_any = overview_df[(overview_df['type'] == 'Built-up') & (overview_df['what'] == 'Proportion NTL')]['value'].values[0]
    built_median_any = overview_df[(overview_df['type'] == 'Built-up') & (overview_df['what'] == 'Median NTL value for any NTL')]['value'].values[0]
    built_median_any = float(built_median_any) / 100
    
    small_perc_mon = overview_df[(overview_df['type'] == 'Small settlements') & (overview_df['what'] == 'Proportion monitorable')]['value'].values[0]
    small_perc_any = overview_df[(overview_df['type'] == 'Small settlements') & (overview_df['what'] == 'Proportion NTL')]['value'].values[0]
    small_median_any = overview_df[(overview_df['type'] == 'Small settlements') & (overview_df['what'] == 'Median NTL value for any NTL')]['value'].values[0]
    small_median_any = float(small_median_any) / 100
    
    
    # Create a new DataFrame with the new row

    new_row_df_figure4 = pd.DataFrame({
        'countryname': [countryname],
        'built-perc_mon': [built_perc_mon],
        'built-perc_any': [built_perc_any],
        'built_median_any': [built_median_any],
        'small-perc_mon': [small_perc_mon],
        'small_perc_any': [small_perc_any],
        'small_median_any': [small_median_any]
    })
    
    # Concatenate the new row DataFrame with the existing DataFrame
    df_figure4 = pd.concat([df_figure4, new_row_df_figure4], ignore_index=True)

# Save the dataframe

df_figure4.to_csv('stats_reporting_figure4.csv', index=False)
print("Overview created. Stats for figure 4 are in files.")
